More cleanup
And refactoring
Working on #22 and #23
