+++ tests/.flake8
@@ -1,2 +1,3 @@
 [flake8]
 max-line-length = 1000000
+ignore = F811
