+++ tests/utils_for_testing.py
@@ -1,0 +1,38 @@
+import pytest
+from json import loads, dumps
+
+
+class ClientSessionMocker:
+    def __init__(self, text, status):
+        self._text = text
+        self.status = status
+
+    async def text(self):
+        return self._text
+
+    async def json(self):
+        return loads(self._text)
+
+    async def __aexit__(self, exc_type, exc, tb):
+        pass
+
+    async def __aenter__(self):
+        return self
+
+
+@pytest.fixture()
+def good_transform_request(mocker):
+    '''
+    Setup to run a good transform request that returns a single file.
+    '''
+    called_json_data = {}
+
+    def call_post(data_dict_to_save: dict, json=None):
+        data_dict_to_save.update(json)
+        return ClientSessionMocker(dumps({"request_id": "1234-4433-111-34-22-444"}), 200)
+    mocker.patch('aiohttp.ClientSession.post', side_effect=lambda _, json: call_post(called_json_data, json=json))
+
+    r2 = ClientSessionMocker(dumps({"files-remaining": "0", "files-processed": "1"}), 200)
+    mocker.patch('aiohttp.ClientSession.get', return_value=r2)
+
+    return called_json_data
